requirements.txt
backports.functools-lru-cache==1.4
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
crayons==0.1.2
future==0.16.0
idna==2.6
Pillow==5.0.0
requests==2.18.4
selene==1.0.0a13
selenium==3.8.1
six==1.11.0
urllib3==1.22
webdriver-manager==1.5.1
